# Conflicts:
#	app/src/main/java/com/red_velvet/yumhub/domain/models/recipes/SearchRecipeEntity.kt
#	app/src/main/java/com/red_velvet/yumhub/ui/signUp/SignUpViewModel.kt
#	app/src/main/res/layout/fragment_home.xml
#	app/src/main/res/navigation/nav_graph.xml

+++ app/src/main/java/com/red_velvet/yumhub/domain/mapper/MealPlanDtoMapper.kt
@@ -8,7 +8,7 @@ import com.red_velvet.yumhub.remote.resources.meal_plan.ValueResource
 
 fun MealPlanLocalDto.toMealPlanDto(): AddMealResource {
     return AddMealResource(
-        date = timestamp.toInt().orZero(),
+        date = timestamp.orZero(),
         position = position.orZero(),
         slot = slot.orZero(),
         type = type.orEmpty(),
